backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
itsdangerous==2.2.0
python-dotenv==1.1.1
google-cloud-firestore==2.21.0
gunicorn==23.0.0

